requirements.txt
bs4~=0.0.1
ao3_api
progress~=1.5

requests~=2.25.0
beautifulsoup4~=4.9.3